requirements.txt
djangorestframework==3.13.1
Django==4.0.4
celery[redis]==5.2.6
django-celery-results==2.3.1
whitenoise==6.0.0
